Use 16KB instead of 4MB.
The usbfs buffer memory is apparently shared across all usbfs devs.
So we should try not to use a huge amount of it. Using a 16KB buffer
here has no observable impact on performance as the trasnfer will be
split up in to packets at least that small anyway.

This also removes the abort counter. That doesn't seem to do anything
except make the transfer fail if it has to be split into more than
20 parts. There is no "retry" code to break out of; the loop always
aborts after a single error.