Update EEPROM to latest stable release 2022-01-25
The MSD (and secure-boot variant) bootloader has been updated to
fully initialise SDRAM in order to resolve problems when using
this method to run Linux ramdisks / initrd.
Therefore, MSD boot will not work on a system with a blank or
corrupted EEPROM. If so, an error message will be output to the
UART and the EEPROM should be programmed using the recovery
option.