win32: Update windows installer and make it smaller
Reduce the size of the windows installer by removing the
optional components. It will now just contain the MSD
firmware drive, the new mass-storage gadget (for writing
NVMe) and the recovery.

The optional components including secure-boot-example and
the Raspberry Pi Image can be downloaded from Github
without rebuilding this installer.

Add a helper batch file and shortcut for launching
the mass-storage-gadget.