update-pieeprom: Use system versions of rpi-eeprom tools in APT installs
Modify the PATH variable so that the rpi-eeprom submodule versions
of the helper tools (rpi-eeprom-config) are used if available.
Otherwise, use the scripts from /usr/bin.

This enables the rpiboot package to just state rpi-eeprom as
a dependency instead of duplicating the tools whilst still allowing
git clone versions for custom installs or non Linux systems.