make-boot-image: use mcopy if available
Currently, the script must be run as the root user due to the use of
losetup/mount. That is unnecessary, and sometimes impossible (e.g. in
the context of an automated meta-build system like Yocto).

The mcopy command, typically provided by an "mtools" package, allows
one to write files to and from a FAT filesystem image. Use that if
available, and add a little hint that one could install mtools if the
script is invoked as non-root.