make-boot-image: preserve file attributes
In order to make binary reproducible images, the script must preserve
modification time and other properties of the files as they are
copied. Both when copying out from the given directory and into the
staging directory, and when copying from the staging dir to the final
image.

As FAT does not have the concept of symlinks or hardlinks, we do need
the copy command to follow symlinks, meaning we cannot have -d in
effect, which in turn precludes using the -a option (which is "same as
-dR --preserve=all"). So just add the -p flag, meaning
"--preserve=mode,ownership,timestamps".