Makefile: make it work better in a cross-build scenario
Long ago, the makefile logic made use of the xxd tool to generate a C
array from a binary file. That was removed in commit
768c0563d110 ("Add cross platform method of building in a binary!"),
in favor of the little bin2c helper, presumably because the xxd tool
is not necessarily available everywhere.

However, that then introduced the implicit assumption that ${CC} is a
compiler meant for producing binaries for the current build host, and
thus breaks down in a cross-build setup, such as when trying to
integrate this in a Yocto recipe.

Adapt the logic so that when xxd is available, use that, and if not,
honour a CC_FOR_BUILD environment variable if set, letting that
default to $(CC). This way, a cross-build environment can either make
sure to add build-time dependency on xxd-native, or export an
appropriate CC_FOR_BUILD variable.