secure-boot-example: Update to work on both Compute Module 4 and Compute Module 5
Update the image to 64-bit to work on both BCM2711 and BCM2712
and update the README to use the sign.sh counter-signing wrapper.

Remove references to BETA firmware, these are no longer included
in APT updates. If a binary is provided on a Github Issue then
a note will be added indicating that secure-boot is not supported.