mass-storage-64: Suggest docker build environment
Fixes: #278

Also fixes the path to the expected output image.

Given that I failed the documented build on my host Ubuntu 22.04.5 LTS
system, and the maintainer reported building under the same, it seems
reasonable to suggest the built-in upstream docker image as a more host-
independent build environment that is known to work.