rpiboot: Use mass-storage-gadget-64 by default on BCM2711
Change the 2711 default behaviour to use mass-storage-gadget64
if the directory argument is not specified.
The old msd.elf can still be run by executing -d msd but this is
not recommended because it is much slower than the Linux
mass-storage-gadget. The msd.elf does not and will not
support NVMe / USB block devices.