msd: Remove legacy msd/start4.elf and use a common legacy msd .elf file
The legacy VPU firmware MSD drivers are deprecated on CM4 / Pi4.
Remove this but link to a generic legacy VPU MSD firmware which
supports 2711 and earlier chips.

N.B. mass-storage-gadget64-cm3 implements same functionality as the
Linux mass-storage-gadget. The only difference is in how the files
are loaded because older boards don't support a bootloader ramdisk.

To run the legacy VPU msd firmware on 2711 run
sudo rpiboot -d ./msd