rpi-make-boot-image: Prefer 512-byte clusters
512-byte clusters allow for a more efficient use of space for small
files (such as overlays) at the cost of a larger total number of
clusters (and earlier promotion from FAT16 to FAT32 as image size
increases).

FAT16 can support a maximum of 65,526 clusters meaning the cross-over
point from FAT16 to FAT32 will be ~32MiB for 512-byte clusters.

As FAT32 is limited to a maximum of 268,435,445 clusters,
rpi-make-boot-image will be limited to a maximum disk image size of
~128GiB when using 512-byte clusters.