update-pieeprom.sh: Remove redundant SIGNED_BOOT=1 flag
The SIGNED_BOOT=1 behaviour is set if the OTP bits are set.
Secure-boot flow without actually programming the OTP which makes
it secure was useful for bringup on BCM2711 but just causes
confusion now.

A future change to the 2711 recovery.bin will set revoke_devkey if
program_pubkey=1 is set to avoid the wrong combination of OTP bits
from being set.

N.B It's possible to test boot.img ramdisks without secure-boot.

Fixes: https://github.com/raspberrypi/usbboot/issues/297