make: Remove BUILD_TIMESTAMP
The build-time is not used, just the date. Remove BUILD_TIMESTAMP
and just allow BUILD_DATE to overriden.
This fixes a problem with the previous macOS patch which would
fail if coreutils was installed to brew because -j is not a valid
flag in that version of the date command.