Add a Hardware Security Module wrapper script interface
Add support for HSMs by extracting by adding an interface where
the PKCS#1 v1.5 signature (RSA2048 SHA256) generation can be
implemented by a dedicated helper script.

This approach makes it easier to support multiple HSM APIs
without adding specific knowledge of HSMs to this code.

The algorithm choice is provided by the API but is very
unlikely to change because that would require updating the
BCM2711 / BCM2712 bootroms.

A software implemenation (secure-boot-example/example-hsm-wrapper)
is provided to define the interface and act as reference for
unit tests.