Makefile: add platform-specific default mass-storage-gadget directory
Move the default mass-storage-gadget64 directory path from a hardcoded
value in main.c to a build-time configurable variable in the Makefile.
On Windows, this defaults to the standard installation path used by the
Windows installer. On Unix systems, the behavior remains unchanged.

This allows rpiboot to be invoked without the -d parameter on Windows.

Fixes: https://github.com/raspberrypi/usbboot/issues/342

Signed-off-by: Nicolai Buchwitz <nb@tipi-net.de>